Refactor batch deletion functionality: separate controller logic and enhance file handling
