+++ controllers/index.js
@@ -20,6 +20,6 @@ router.use('/tools', require('./tools'));
 router.use('/ctltools', require('./ctltools'));
 router.use('/storageproviders', require('./storageproviders'));
 router.use('/base64files/', require('./base64files'));
-router.use('/batch-deletion/', require('./batch-deletion'));
+router.use('/batch-deletion/', require('./batch-deletion/batch-deletion'));
 
 module.exports = router;
